## Motivation

I noticed that nethermind's addmod implementation was faster than ours, and implemented differently

## Solution

Using their implementation, which does one overflowing add, then a div_rem if overflowing, I ported it to rust

Benchmarks so far:
```
add_mod/0               time:   [320.13 ps 321.23 ps 322.47 ps]
                        change: [-0.7095% +0.0158% +0.6198%] (p = 0.96 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  9 (9.00%) high mild
  1 (1.00%) high severe

add_mod/64              time:   [18.756 ns 18.825 ns 18.896 ns]
                        change: [+71.888% +76.778% +80.854%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe

add_mod/128             time:   [23.228 ns 23.361 ns 23.508 ns]
                        change: [-52.836% -51.662% -50.226%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) high mild
  4 (4.00%) high severe

add_mod/192             time:   [28.678 ns 28.865 ns 29.057 ns]
                        change: [-50.733% -49.805% -48.754%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

add_mod/256             time:   [30.868 ns 31.042 ns 31.223 ns]
                        change: [-51.260% -49.086% -46.023%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

add_mod/384             time:   [37.132 ns 37.389 ns 37.647 ns]
                        change: [-47.157% -43.354% -36.820%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

add_mod/512             time:   [43.611 ns 44.109 ns 44.642 ns]
                        change: [-45.368% -39.805% -31.381%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

add_mod/4096            time:   [357.04 ns 362.27 ns 367.49 ns]
                        change: [-35.678% -16.946% +7.8972%] (p = 0.15 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
```

Unfortunately we can't just make an array of `BITS + 1` or `LIMBS + 1`, which would be much easier, maybe there's some hack we can do. I also need to figure out why performance is worse in `add_mod/64`

Geth's implementation seems faster (from [gas-cost-estimator](https://github.com/imapp-pl/gas-cost-estimator/blob/master/docs/report_stage_iv.md) benchmarks), impl is here:
https://github.com/holiman/uint256/blob/a17fcfb6f8b0245533928fa96aace2cd4c2e9b18/uint256.go#L217-L283

So I may see if that ends up faster or can be generalized

## PR Checklist

- [ ] Added Tests
- [ ] Added Documentation
- [ ] Updated the changelog

<!-- This template is based on https://github.com/tokio-rs/tokio/blob/tokio-1.13.0/.github/PULL_REQUEST_TEMPLATE.md and https://github.com/gakonst/ethers-rs/blob/0.5.3/.github/PULL_REQUEST_TEMPLATE.md -->