@DaniPopes 2025-05-26
Superseded by https://github.com/recmo/uint/pull/469